# src/Infrastructure/BulletinBoard.WebAPI/wwwroot/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Доска объявлений",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5129/api/"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Создаёт нового пользователя.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "Имя пользователя.",
                    "type": "string",
                    "maxLength": 100
                  },
                  "is_admin": {
                    "description": "Является ли пользователь администратором.",
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "is_admin"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно создан.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserId"
                }
              }
            }
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Позволяет искать по списку пользователей.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Количество результатов поиска, которое необходимо вернуть.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "searchName",
            "in": "query",
            "description": "Поисковый запрос по имени пользователя.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchIsAdmin",
            "in": "query",
            "description": "Если true, то будут выведены пользователи со статусом is_admin = true.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Задаёт порядок сортировки.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "created",
                "name",
                "is_admin"
              ],
              "default": "created"
            }
          },
          {
            "name": "desc",
            "in": "query",
            "description": "При передаче true значения будут возвращены в обратном порядке.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "description": "Вернёт пользователей, добавленных не раньше указанной даты.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "yyyy-MM-ddThh:mm+hh:mm"
            }
          },
          {
            "name": "createdTo",
            "in": "query",
            "description": "Вернёт пользователей, добавленных не позже указанной даты.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "yyyy-MM-ddThh:mm+hh:mm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Возвращает информацию о пользователе.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор пользователя.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Обновляет информацию о пользователе.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор пользователя.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "Имя пользователя.",
                    "type": "string",
                    "maxLength": 100
                  },
                  "is_admin": {
                    "description": "Является ли пользователь администратором.",
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "is_admin"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Информация успешно обновлена."
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Удаляет пользователя.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор пользователя.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Пользователь успешно удалён."
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/bulletins": {
      "post": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Создаёт новое объявление.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "description": "Текст объявления.",
                    "type": "string",
                    "maxLength": 1000
                  },
                  "rating": {
                    "description": "Рейтинг объявления.",
                    "type": "integer"
                  },
                  "expiry": {
                    "description": "Дата истечения объявления.",
                    "type": "string",
                    "format": "yyyy-MM-ddThh:mm+hh:mm"
                  },
                  "userId": {
                    "description": "Идентификатор пользователя - автора объявления.",
                    "type": "string",
                    "format": "uuid"
                  },
                  "image": {
                    "description": "Изображение.",
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "text",
                  "rating",
                  "expiry",
                  "userId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Объявление успешно создано.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinId"
                }
              }
            }
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Позволяет искать по списку объявлений.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Количество результатов поиска, которое необходимо вернуть.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "searchNumber",
            "in": "query",
            "description": "Поисковый запрос по номеру объявления.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchText",
            "in": "query",
            "description": "Поисковый запрос по тексту объявления.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchUserId",
            "in": "query",
            "description": "Поисковый запрос по идентификатору пользователя - автора объявления.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Задаёт порядок сортировки.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "created",
                "number",
                "text",
                "rating"
              ],
              "default": "created"
            }
          },
          {
            "name": "desc",
            "in": "query",
            "description": "При передаче true значения будут возвращены в обратном порядке.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "ratingFrom",
            "in": "query",
            "description": "Вернёт объявления с рейтингом не ниже переданного.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "ratingTo",
            "in": "query",
            "description": "Вернёт объявления с рейтингом не выше переданного.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "description": "Вернёт объявления с датой создания не раньше переданной даты.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "yyyy-MM-ddThh:mm+hh:mm"
            }
          },
          {
            "name": "createdTo",
            "in": "query",
            "description": "Вернёт объявления с датой создания не позже переданной даты.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "yyyy-MM-ddThh:mm+hh:mm"
            }
          },
          {
            "name": "expiryFrom",
            "in": "query",
            "description": "Вернёт объявления с датой истечения не раньше переданной даты.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "yyyy-MM-ddThh:mm+hh:mm"
            }
          },
          {
            "name": "createdTo",
            "in": "query",
            "description": "Вернёт объявления с датой истечения не позже переданной даты.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "yyyy-MM-ddThh:mm+hh:mm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bulletin"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/bulletins/{id}": {
      "get": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Возвращает информацию об объявлении.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор объявления.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bulletin"
                }
              }
            }
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Объявление не найдено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Обновляет информацию об объявлении.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор объявления.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "description": "Текст объявления.",
                    "type": "string",
                    "maxLength": 1000
                  },
                  "rating": {
                    "description": "Рейтинг объявления.",
                    "type": "integer"
                  },
                  "expiry": {
                    "description": "Дата истечения объявления.",
                    "type": "string",
                    "format": "yyyy-MM-ddThh:mm+hh:mm"
                  },
                  "image": {
                    "description": "Изображение.",
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "text",
                  "rating",
                  "expiry"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Информация успешно обновлена."
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Объявление не найдено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Удаляет объявление.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор объявления.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Объявление успешно удалёно."
          },
          "400": {
            "description": "Параметры запроса переданы неверно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Объявление не найдено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Сообщение об ошибке."
          },
          "status": {
            "type": "integer",
            "description": "Статус-код ошибки."
          },
          "detail": {
            "type": "string",
            "description": "Подробное сообщение об ошибке."
          }
        }
      },
      "UserId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Идентификатор пользователя."
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Идентификатор пользователя."
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя."
          },
          "is_admin": {
            "type": "boolean",
            "description": "Является ли пользователь администратором."
          },
          "created_utc": {
            "type": "string",
            "format": "date-time",
            "description": "Дата и время создания пользователя."
          }
        }
      },
      "BulletinId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Идентификатор объявления."
          }
        }
      },
      "Bulletin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Идентификатор объявления."
          },
          "number": {
            "type": "integer",
            "description": "Порядковый номер объявления."
          },
          "text": {
            "type": "string",
            "description": "Текст объявления."
          },
          "rating": {
            "type": "integer",
            "description": "Рейтинг объявления."
          },
          "created_utc": {
            "type": "string",
            "format": "date-time",
            "description": "Дата и время создания объявления."
          },
          "expiry_utc": {
            "type": "string",
            "format": "date-time",
            "description": "Дата и время истечения объявления."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "Идентификатор пользователя - автора объявления."
          },
          "image_preview": {
            "type": "string",
            "nullable": true,
            "description": "Ссылка на превью изображения."
          },
          "image_full": {
            "type": "string",
            "nullable": true,
            "description": "Ссылка на изображение в полном размере."
          }
        }
      }
    }
  }
}